requirements.txt
tensorflow==2.20.0
streamlit==1.30.0
numpy==1.26.4
pillow==10.4.0
h5py==3.14.0
